requirements.txt
colorama>=0.4.5
requests>=2.28.1
python-dotenv>=0.20.0
toml>=0.10.2